end Task

+++ AMPMI/WebSite.EndPoint/Areas/Admin/Controllers/UserController.cs
@@ -1,4 +1,6 @@
-﻿using AQS_Application.Interfaces.IServices.BaseServices;
+﻿using AQS_Application.Dtos.IdentityServiceDto;
+using AQS_Application.Interfaces.IServices.BaseServices;
+using AQS_Application.Interfaces.IServices.IdentityServices;
 using AQS_Common.Enums;
 using Microsoft.AspNetCore.Mvc;
 using WebSite.EndPoint.Areas.Admin.Models.Company;
@@ -12,12 +14,15 @@ namespace WebSite.EndPoint.Areas.Admin.Controllers
     {
         private readonly ICompanyService _companyService;
         private readonly IFileServices _fileServices;
+        private readonly IRegistrationService _registrationService;
         private const string PictureFolder = "CompanyProfile";
+        private const string Role = "Company";
 
-        public UserController(ICompanyService companyService, IFileServices fileServices)
+        public UserController(ICompanyService companyService, IFileServices fileServices, IRegistrationService registeredServices)
         {
             _companyService = companyService;
             _fileServices = fileServices;
+            _registrationService = registeredServices;
         }
         public async Task<IActionResult> UserList()
         {
@@ -37,6 +42,12 @@ namespace WebSite.EndPoint.Areas.Admin.Controllers
         }
         public async Task<IActionResult> DeleteUser(long id)
         {
+            var resultReg = await _registrationService.DeleteUserAsync(id);
+            if (!resultReg)
+            {
+                TempData["msg"] = "خطا در هنگام حذف کاربر";
+                return RedirectToAction(nameof(UserList));
+            }
             var result = await _companyService.Delete(id);
             if (result == ResultOutPutMethodEnum.savechanged)
                 TempData["msg"] = "حذف با موفقیت انجام شد";
@@ -118,7 +129,7 @@ namespace WebSite.EndPoint.Areas.Admin.Controllers
 
                 var dto = CompanyVM.MapToDto(CompanyVM);
 
-                await _companyService.IsCompany(dto.Id , dto.IsCompany);
+                await _companyService.IsCompany(dto.Id, dto.IsCompany);
 
                 var resultMessage = await _companyService.UpdateEditProfile(dto);
                 if (resultMessage == ResultOutPutMethodEnum.savechanged || resultMessage == ResultOutPutMethodEnum.dontSaved)
@@ -136,6 +147,72 @@ namespace WebSite.EndPoint.Areas.Admin.Controllers
                 return View(CompanyVM);
             }
         }
+        public async Task<IActionResult> CreateUser()
+        {
+            return View("EditUser", new CompanyVM());
+        }
+        [HttpPost]
+        public async Task<IActionResult> CreateUser(CompanyVM companyVM)
+        {
+            try
+            {
+                RegisterIdentityDTO registerIdentityDTO = new()
+                {
+
+                    Mobile = companyVM.MobileNumber,
+                    CompanyName = companyVM.Name,
+                    ManagerName = companyVM.ManagerName,
+                    Email = companyVM.Email,
+                    Password = companyVM.MobileNumber,
+                    Address = companyVM.Address
+                };
+
+                var createNew = await _registrationService.RegisterAsync(registerIdentityDTO, Role, false);
+                if (createNew.userId <= 0)
+                {
+                    ViewData["error"] = createNew.errorMessage;
+                    return RedirectToAction(nameof(EditUser), new { id = createNew.userId });
+                }
+                if (companyVM.IsLogoChanged)
+                {
+                    if (companyVM.Logo != null)
+                    {
+                        string newLogo = await AddPicture(companyVM.Logo);
+                        if (string.IsNullOrEmpty(newLogo))
+                        {
+                            ViewData["error"] = "خطایی در هنگام دخیره لوگو رخ داد";
+                            return RedirectToAction(nameof(EditUser), new { id = createNew.userId });
+                        }
+                        companyVM.LogoRout = newLogo;
+                    }
+                }
+                if (companyVM.IsBannerChanged)
+                {
+                    if (companyVM.Banner != null)
+                    {
+                        string newBanner = await AddPicture(companyVM.Banner);
+                        if (string.IsNullOrEmpty(newBanner))
+                        {
+                            ViewData["error"] = "خطایی در هنگام دخیره بنر رخ داد";
+                            return RedirectToAction(nameof(EditUser), new { id = createNew.userId });
+                        }
+                        companyVM.BannerRout = newBanner;
+                    }
+                }
+
+                var dto = companyVM.MapToDto(companyVM);
+
+                await _companyService.IsCompany(dto.Id, dto.IsCompany);
+
+
+                return RedirectToAction(nameof(EditUser), new { id = createNew.userId });
+            }
+            catch (Exception ex)
+            {
+                ViewData["error"] = ex.Message;
+                return RedirectToAction(nameof(EditUser), new { id = 0 });
+            }
+        }
         private async Task<bool> DeletePicture(string route)
         {
             return await _fileServices.DeleteFile(route);
@@ -143,6 +220,14 @@ namespace WebSite.EndPoint.Areas.Admin.Controllers
         private async Task<string> AddPicture(IFormFile newPicture)
         {
             return await _fileServices.SaveFileAsync(newPicture, PictureFolder);
+        }
+        [HttpPost]
+        public async Task<IActionResult> Save(CompanyVM companyVM)
+        {
+            if (companyVM.Id > 0)
+                return await EditUser(companyVM);
+            else
+                return await CreateUser(companyVM);
         }
     }
 }
